+++ devnotes/toolkit/CHANGELOG.md
@@ -1,3 +1,4 @@
 # Toolkit CHANGELOG
 
-(아직 변경 이력 없음)
+## 2026-09-19 (97차)
+- lead_decel/ 추가: parse_lead_log.py, merge_lead_series.py, counterfactual.py, mpc_replica.py, closed_loop.py (선행차 감속 과민 반응 분석용, README 참고)